requirements.txt
pandas
requests
sqlalchemy
scikit-learn
python-dotenv
jupyter
matplotlib
seaborn
psycopg2-binary
google-cloud-bigquery
google-auth
google-auth-oauthlib
google-auth-httplib2
